move test w/ burnettk
