+++ spiffworkflow-backend/tests/spiffworkflow_backend/unit/test_user_service.py
@@ -1,8 +1,9 @@
-"""Process Model."""
-
 from flask.app import Flask
 from starlette.testclient import TestClient
 
+from spiffworkflow_backend.models.db import db
+from spiffworkflow_backend.models.human_task import HumanTaskModel
+from spiffworkflow_backend.models.process_instance import ProcessInstanceModel
 from spiffworkflow_backend.services.user_service import UserService
 from tests.spiffworkflow_backend.helpers.base_test import BaseTest
 
@@ -40,3 +41,39 @@ class TestUserService(BaseTest):
         everybody_group = UserService.find_or_create_group("everybodyGroup")
         UserService.add_waiting_group_assignment("REGEX:.*", everybody_group)
         assert initiator_user.groups[0] == everybody_group
+
+    def test_newly_added_user_to_group_gets_access_to_existing_tasks(
+        self,
+        app: Flask,
+        client: TestClient,
+        with_db_and_bpmn_file_cleanup: None,
+    ) -> None:
+        user1 = self.find_or_create_user("user1")
+        awesome_group = UserService.find_or_create_group("Awesome")
+        process_instance = ProcessInstanceModel(
+            process_initiator=user1,
+            process_model_identifier="test/process",
+            process_model_display_name="Test Process",
+        )
+
+        db.session.add(process_instance)
+        db.session.commit()
+
+        human_task = HumanTaskModel(
+            process_instance_id=process_instance.id,
+            task_id="task_1",
+            task_name="Task 1",
+            task_type="UserTask",
+            task_status="READY",
+            lane_assignment_id=awesome_group.id,
+        )
+        db.session.add(human_task)
+        db.session.commit()
+
+        assert user1 not in human_task.potential_owners
+        UserService.add_user_to_group(user1, awesome_group)
+
+        # We need to refresh the human_task object or its potential_owners relationship
+        db.session.refresh(human_task)
+
+        assert user1 in human_task.potential_owners
